requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.4
pytest>=8.1.1
httpx>=0.27.0
requests>=2.31.0
lxml>=5.1.0
xmlschema>=3.0.1
pyyaml>=6.0.1
slowapi>=0.1.9
structlog>=24.1.0
prometheus-client>=0.20.0
sqlalchemy>=2.0.30
alembic>=1.13.0
pydantic-settings>=2.2.0
redis>=5.0.4
psycopg2-binary>=2.9.9
openpyxl>=3.1.0
pandas>=2.2.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
